Adding form on basic info screen
I've added the form layout on the basic info screen. This is hard coded and will need to be dynamically built. Then I can place the text boxes into the form, and then work on the groupings of the correct items to display at the same time, then comes the logic to show which groupings are required.