textboxes populate based on user input
Textboxes start as invisible and now they appear when it is appropriate that they appear. Next item for review is making the textboxes appear only when appropriate. Currently any answer prompts the textboxes to appear, but you may actually only want the text box to appear if a "yes" or a "no" is answered.